add hints (still a bit buggy)
Actually it's not a bug as designed - even though it looks like
you only have one letter left in a word like "fini_hed", typing
"s" wont do anything, as the combo logic is still looking for
"finis" to display the "s".  Also, the "s" won't lock in until
you finish spelling the whole word, even though it's already
displayed as complete.

The solution requires modifying the combo matching logic to
consider if any letters in a matching word are already filled in.

I'll make those adjustments in the next commit.